pyproject.toml
[tool.poetry]
name = "oreilly-discord-bot"
version = "0.1.0"
description = "Discord Bot to Notification O'Reilly Book News."
authors = ["HEKUCHAN <heitorhirose@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.3.2"
alembic = "^1.13.1"
sqlalchemy = "^2.0.28"
sqlalchemy-utils = "^0.41.1"
pydantic = {extras = ["dotenv"], version = "^2.6.4"}
pydantic-settings = "^2.2.1"
loguru = "^0.7.2"
aiosqlite = "^0.20.0"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
ruff = "^0.3.2"
nox = "^2024.3.2"
pytest = "^8.1.1"
dpytest = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
